refactor(split-button): rename subcomponent .scss files to .module.scss
