+++ packages/split-button/src/SplitButton/components/ActionButton/ActionButton.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import classnames from "classnames"
 import { BaseButton, BaseButtonProps } from "../BaseButton"
-import styles from "./ActionButton.scss"
+import styles from "./ActionButton.module.scss"
 
 export type ActionButtonProps = Omit<BaseButtonProps, "icon">
